before first correct
