+++ src/App.tsx
@@ -30,4 +30,3 @@ function App() {
 }
 
 export default App;
-
