--- src/components/Error/Error.tsx
@@ -1,3 +1,0 @@
-export const Error = () => {
-  return <div data-testid="courses-loading">Error!</div>;
-};
