--- src/components/Error/index.ts
@@ -1,1 +1,0 @@
-export * from './Error';
