--- src/components/Lesson/index.ts
@@ -1,1 +1,0 @@
-export * from './Lesson';
