--- src/components/Loader/Loader.tsx
@@ -1,3 +1,0 @@
-export const Loader = () => {
-  return <div data-testid="courses-loading">Loading...</div>;
-};
