--- src/components/Loader/index.ts
@@ -1,1 +1,0 @@
-export * from './Loader';
