+++ src/components/index.ts
@@ -1,8 +1,3 @@
-import { CourseCard } from './CourseCard';
-import { Lesson } from './Lesson';
-import { Pagination } from './Pagination';
-import { VideoPlayer } from './VideoPlayer';
-import { Loader } from './Loader';
-import { Error } from './Error';
+import { CourseCard } from "./CourseCard";
 
-export { CourseCard, Lesson, Pagination, VideoPlayer, Loader, Error };
+export { CourseCard };
