+++ src/utils/constants.ts
@@ -1,26 +1,22 @@
-export const URL = 'https://api.wisey.app';
+export const URL = "https://api.wisey.app";
 
 export const API_VERSIONS = {
-  V1: 'api/v1',
+    V1: "api/v1",
 };
 
 export const STORAGE_KEYS = {
-  TOKEN: 'TOKEN',
+    TOKEN: "TOKEN",
 };
 
 export const ROUTER_KEYS = {
-  COURSES: 'courses',
-  COURSE: 'course',
+    COURSES: "courses",
+    COURSE: "course",
 };
 
 export const QUERY_KEYS = {
-  ROOT: '/',
-  AUTH: 'auth',
-  ANONYMOUS: 'anonymous',
-  CORE: 'core',
-  PREVIEW_COURSES: 'preview-courses',
+    ROOT: "/",
+    AUTH: "auth",
+    ANONYMOUS: "anonymous",
+    CORE: "core",
+    PREVIEW_COURSES: "preview-courses",
 };
-
-export const COURSES_PER_PAGE = 10;
-
-export const AUTO_CLOSE_TOAST = 2000;
